refactors and puts back the original Aggregations component
